move graphql module to new module package
